streamlit_app/requirements.txt
pandas
matplotlib
ipython
numpy
seaborn
nltk
scikit-learn==1.1.3
scikit-learn-extra
gensim
sacrebleu
pyspellchecker
spacy==3.6.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0.tar.gz 
https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.6.0/fr_core_news_sm-3.6.0.tar.gz 
pillow
wordcloud
pillow
streamlit
pathlib
gensim
networkx==2.7.0
transformers
tensorflow
sentencepiece
openai-whisper
sounddevice==0.4.6
xformers
translate
speechrecognition
audio_recorder_streamlit
wave
whisper
wavio
filesplit
regex
pydot
graphviz
gtts
extra_streamlit_component